policy, rpc: Disallow non-0 mod fees
Reject ephemeral dust entry into mempool
if base or modified fees are non-0.

prioritisetransaction throws an error
if called on a transaction in its mempool
that has dust.

This has the effect that unspent dust should
never be mined unless minrelay et al are changed
to 0 fee, or the operator simply redefines that dust
values using `-dustrelayfee` to make dust threshholds
lower.